Restructuring and reworking the project entirely:
- Added interface implementation to support multiple different l2j
servers
- Separated the code from the original l2j project.
