+++ .gitignore
@@ -1,0 +1,22 @@
+# Eclipse
+*.classpath
+*.project
+*.settings
+*.metadata
+bin/
+
+# Idea
+*.iml
+*/.idea/
+
+# Gradle
+/.gradle/
+*/build/
+
+
+.DS_Store
+*.log
+
+# Cache folders
+*/.cache/
+build
