+++ VotingRewardAPI/build.gradle
@@ -1,0 +1,40 @@
+apply plugin: 'java'
+apply plugin: 'distribution'
+
+defaultTasks('distZip')
+
+tasks.withType(JavaCompile) {
+	options.encoding = 'UTF-8'
+}
+
+dependencies {
+	compile project(':VotingRewardInterfaceProvider')
+}
+
+def generalManifest = manifest {
+	attributes('Built-By': System.getProperty('user.name'),
+			'Implementation-URL': 'https://github.com/UnAfraid/VotingRewardAPI')
+}
+
+
+jar {
+	archiveName = 'VotingRewardAPI.jar'
+	manifest {
+		from(generalManifest)
+	}
+}
+
+
+distributions {
+	main {
+		contents {
+			from('dist')
+			into('libs') {
+				from(configurations.runtime)
+				from(jar)
+			}
+		}
+	}
+}
+
+
