+++ VotingRewardAPI/src/main/java/com/github/unafraid/votingreward/model/UserScope.java
@@ -1,0 +1,74 @@
+/*
+ * Copyright (C) 2014-2015 Vote Rewarding System
+ * 
+ * This file is part of Vote Rewarding System.
+ * 
+ * Vote Rewarding System is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Vote Rewarding System is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.github.unafraid.votingreward.model;
+
+import com.github.unafraid.votingreward.interfaceprovider.api.IPlayerInstance;
+
+public enum UserScope
+{
+	ACCOUNT
+	{
+		@Override
+		public String getData(IPlayerInstance player)
+		{
+			return player.getAccountName();
+		}
+	},
+	IP
+	{
+		@Override
+		public String getData(IPlayerInstance player)
+		{
+			return player.getIPAddress();
+		}
+	},
+	HWID
+	{
+		@Override
+		public String getData(IPlayerInstance player)
+		{
+			return player.getHWID();
+		}
+		
+		@Override
+		public boolean isSupported(IPlayerInstance player)
+		{
+			return player.isHWIDSupported();
+		}
+	};
+	
+	public abstract String getData(IPlayerInstance player);
+	
+	public boolean isSupported(IPlayerInstance player)
+	{
+		return true;
+	}
+	
+	public static UserScope findByName(String name)
+	{
+		for (UserScope scope : values())
+		{
+			if (scope.name().equals(name))
+			{
+				return scope;
+			}
+		}
+		return null;
+	}
+}
