+++ VotingRewardInterfaceProvider/build.gradle
@@ -1,0 +1,36 @@
+apply plugin: 'java'
+apply plugin: 'distribution'
+
+defaultTasks('distZip')
+
+tasks.withType(JavaCompile) {
+	options.encoding = 'UTF-8'
+}
+
+def generalManifest = manifest {
+	attributes('Built-By': System.getProperty('user.name'),
+			'Implementation-URL': 'https://github.com/UnAfraid/VotingRewardAPI')
+}
+
+
+jar {
+	archiveName = 'VotingRewardInterfaceProvider.jar'
+	manifest {
+		from(generalManifest)
+	}
+}
+
+
+distributions {
+	main {
+		contents {
+			from('dist')
+			into('libs') {
+				from(configurations.runtime)
+				from(jar)
+			}
+		}
+	}
+}
+
+
