+++ VotingRewardInterfaceProvider/src/main/java/com/github/unafraid/votingreward/interfaceprovider/api/IOnVoicedCommandHandler.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright (C) 2014-2015 Vote Rewarding System
+ * 
+ * This file is part of Vote Rewarding System.
+ * 
+ * Vote Rewarding System is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Vote Rewarding System is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.github.unafraid.votingreward.interfaceprovider.api;
+
+/**
+ * @author UnAfraid
+ */
+public interface IOnVoicedCommandHandler
+{
+	public boolean useVoicedCommand(String command, IPlayerInstance player, String params);
+	
+	public String[] getVoicedCommandList();
+}
