+++ VotingRewardInterfaceProvider/src/main/java/com/github/unafraid/votingreward/interfaceprovider/api/IPlayerInstance.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright (C) 2014-2015 Vote Rewarding System
+ * 
+ * This file is part of Vote Rewarding System.
+ * 
+ * Vote Rewarding System is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Vote Rewarding System is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.github.unafraid.votingreward.interfaceprovider.api;
+
+import com.github.unafraid.votingreward.interfaceprovider.model.RewardItemHolder;
+
+/**
+ * @author UnAfraid
+ */
+public interface IPlayerInstance
+{
+	public boolean isGM();
+	
+	public void sendMessage(String message);
+	
+	public void addItem(RewardItemHolder holder);
+	
+	public String getAccountName();
+	
+	public String getIPAddress();
+	
+	public String getHWID();
+	
+	public boolean isHWIDSupported();
+}
