+++ VotingRewardL2jInterface/build.gradle
@@ -1,0 +1,41 @@
+apply plugin: 'java'
+apply plugin: 'distribution'
+
+defaultTasks('distZip')
+
+tasks.withType(JavaCompile) {
+	options.encoding = 'UTF-8'
+}
+
+dependencies {
+	compile project(':VotingRewardInterfaceProvider')
+	compile 'org.slf4j:slf4j-api:1.7.12'
+	compile 'org.slf4j:slf4j-jdk14:1.7.12'
+	compile(fileTree(dir: 'libs', include: '*.jar'))
+}
+
+def generalManifest = manifest {
+	attributes('Built-By': System.getProperty('user.name'),
+			'Implementation-URL': 'https://github.com/UnAfraid/VotingRewardAPI')
+}
+
+jar {
+	archiveName = 'VotingRewardL2JInterface.jar'
+	manifest {
+		from(generalManifest)
+	}
+}
+
+
+distributions {
+	main {
+		contents {
+			from('dist')
+			into('libs') {
+				from(jar)
+			}
+		}
+	}
+}
+
+
