+++ VotingRewardL2jInterface/src/main/resources/META-INF/services/com.github.unafraid.votingreward.interfaceprovider.IVotingRewardInterfaceProvider
@@ -1,0 +1,1 @@
+com.github.unafraid.votingreward.l2j.L2JInterfaceProvider
