+++ eclipse-settings/org.eclipse.wst.sse.core.prefs
@@ -1,0 +1,3 @@
+eclipse.preferences.version=1
+task-tags/enabled=true
+task-tags/use-project-settings=true
