+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Fri Jul 17 18:52:11 EEST 2015
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.7-bin.zip
